requirements.txt
transformers
requests
torch
uvicorn
logging
fastapi
pydantic
httpx
python-dotenv
bitsandbytes
accelerate>=0.26.0